requirements.txt
aiogram
playwright
rapidfuzz
aiohttp
python-dotenv